requirements.txt
flask==3.1.2
pillow==12.0.0
opencv-python==4.11.0.86
numpy==2.3.4
exifread==3.5.1
reportlab==4.4.4
scikit-image==0.25.2
matplotlib==3.10.7
